# docs/swagger.yaml
definitions:
  entities.CreateUserReq:
    properties:
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      nick_name:
        type: string
      pswd_hash:
        type: string
    type: object
  entities.ErrResp:
    properties:
      errorMessage:
        example: A horrible, terrible, absolutely awful error
        type: string
    type: object
  entities.UserResp:
    properties:
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      pswd_hash:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
info:
  contact:
    name: TomasCruz
    url: https://github.com/TomasCruz/users
  description: Users service
  license:
    name: MIT
    url: https://mit-license.org/
  title: Users
  version: "1.0"
paths:
  /health:
    get:
      operationId: health
      produces:
      - application/json
      responses:
        "204":
          description: Healthy
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.ErrResp'
      summary: health check
  /users:
    get:
      description: list user details
      operationId: list-user
      parameters:
      - description: Country
        in: query
        name: country
        type: string
      - description: Page size
        in: query
        name: page-size
        type: string
      - description: Page number
        in: query
        name: page-number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User detail list
          schema:
            items:
              $ref: '#/definitions/entities.UserResp'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "424":
          description: Database Error
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.ErrResp'
      summary: list users
    put:
      description: creates user
      operationId: create-user
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.CreateUserReq'
      produces:
      - application/json
      responses:
        "201":
          description: User details
          schema:
            $ref: '#/definitions/entities.UserResp'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "409":
          description: Existing email
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "424":
          description: Database or Kafka Error
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.ErrResp'
      summary: creates user
  /users/{user-id}:
    get:
      description: gets user details
      operationId: get-user
      parameters:
      - description: User id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/entities.UserResp'
        "400":
          description: Bad ID
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "424":
          description: Database Error
          schema:
            $ref: '#/definitions/entities.ErrResp'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.ErrResp'
      summary: gets user
swagger: "2.0"